requirements.txt
pillow
requests
numpy
face_recognition
fastapi
pydantic
uvicorn
pytz
hypercorn